moved nav state to parent for locking scroll
