+++ components/NavigationBar.tsx
@@ -1,52 +1,20 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import Logo from "./Logo";
 
-function debounce(fn: Function, ms: number) {
-  let timer: any;
-  function setDebounce() {
-    timer = null;
-    fn.apply(this, arguments);
-  }
-  return () => {
-    clearTimeout(timer);
-    timer = setTimeout(setDebounce, ms);
-  };
-}
-
-const NavigationBar = () => {
+const NavigationBar: React.FC<{
+  isMobile: boolean;
+  isHamburgerOpen: boolean;
+  setIsHamburgerOpen: Function;
+}> = ({ isMobile, isHamburgerOpen, setIsHamburgerOpen }) => {
   const { asPath } = useRouter();
-  const [windowWidth, setWindowWidth] = useState<number | null>(null);
-  const [isHamburgerOpen, setIsHamburgerOpen] = useState<boolean>(false);
-  const [isMobile, setIsMobile] = useState<boolean>(false);
-
-  // track window width in state
-  useEffect(() => {
-    const debounceHandleResize = debounce(() => {
-      setWindowWidth(window.innerWidth);
-    }, 20);
-    window.addEventListener("resize", debounceHandleResize);
-    return () => {
-      window.removeEventListener("resize", debounceHandleResize);
-    };
-  }, []);
-
-  // handle change from mobile to desktop + resetting state
-  useEffect(() => {
-    if (windowWidth > 768) {
-      setIsMobile(false);
-      if (isHamburgerOpen) setIsHamburgerOpen(false);
-    } else {
-      if (!isMobile) setIsMobile(true);
-    }
-  }, [windowWidth]);
 
   return (
     <nav
-      className={`flex justify-center text-lg py-2 bg-gray font-black tracking-wide min-h-[46px] ${
-        isMobile && isHamburgerOpen && "absolute h-full w-full items-start"
-      }`}
+      className={`fixed w-full z-50 flex justify-center text-lg py-2 bg-gray font-black tracking-wide min-h-[46px] ${
+        !isHamburgerOpen && "border-b-4 border-black"
+      } ${isMobile && isHamburgerOpen && "absolute h-full w-full items-start"}`}
     >
       <div
         className={`flex mx-8 ${isMobile && "w-full"} ${
